requirements.txt
# ============================================================
# Meeting AI Assistant — Requirements
# ============================================================

# --- FastAPI Backend ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2

# --- ASR ---
faster-whisper==1.0.3
numpy==1.26.4

# --- Audio capture (optionnel, pour capture micro directe) ---
sounddevice==0.5.1
websocket-client==1.8.0

# --- Streamlit Frontend ---
streamlit==1.40.0
plotly==5.24.1
pandas==2.2.3
altair==5.4.1

# --- Export ---
reportlab==4.2.5       # PDF (optionnel)
markdown==3.7

# --- Google Calendar (optionnel) ---
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# --- Utils ---
python-dotenv==1.0.1
